requirements.txt
Flask
Pillow
num2words
gunicorn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
openpyxl

